V2/requirements.txt
# Core dependencies
git+https://github.com/The-CJ/osu_irc.py.git#egg=osu_irc
asyncpg==0.28.0
aiohttp==3.10.0
aiosqlite==0.19.0
aiomysql==0.2.0
pyyaml==6.0.1
requests==2.32.2
pillow<11.0.0
setuptools>=65.5.1,<70.0.0

# Development dependencies
black==23.7.0
mypy==1.5.1
pytest==7.4.0
pytest-asyncio==0.21.1
